InteractiveChatbot/backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
groq==0.20.0
markdown2==2.5.3
matplotlib==3.10.1
nltk==3.9.1
pandas==2.2.3
protobuf==3.20.3
pymongo==4.6.3
python-dotenv==1.1.0
reportlab==4.3.1
Requests==2.32.3
seaborn==0.13.2
weasyprint==65.0
python-dotenv==1.1.0